serverless/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
boto3==1.34.102
    # via -r requirements.in
botocore==1.34.162
    # via
    #   boto3
    #   s3transfer
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
pydantic==2.9.0
    # via -r requirements.in
pydantic-core==2.23.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
python-ulid==2.4.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via python-dateutil
structlog==22.1.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via pydantic
urllib3==2.3.0
    # via
    #   botocore
    #   requests
